% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_wonders_app}
\alias{run_wonders_app}
\title{Run the Seven Wonders Shiny App}
\usage{
run_wonders_app()
}
\description{
Launches an interactive Shiny application to explore the New 7 Wonders of
the World. Users can filter wonders by continent, structure type, and
construction year
}
\examples{
\donttest{
run_wonders_app()
}
}
